+++ src/main/java/net/helpscout/api/model/Conversation.java
@@ -1,8 +1,5 @@
 package net.helpscout.api.model;
 
-import java.util.Calendar;
-import java.util.List;
-
 import net.helpscout.api.cbo.Status;
 import net.helpscout.api.model.ref.CustomerRef;
 import net.helpscout.api.model.ref.MailboxRef;
@@ -10,13 +7,16 @@ import net.helpscout.api.model.ref.PersonRef;
 import net.helpscout.api.model.ref.UserRef;
 import net.helpscout.api.model.thread.LineItem;
 
+import java.util.Calendar;
+import java.util.List;
+
 public class Conversation {
-	private int id;
-	private int folderId;
+	private Long id;
+	private Long folderId;
 	private boolean isDraft;
-	private int number;
+	private Long number;
 	private Source source;
-	
+
 	private UserRef owner;
 	private MailboxRef mailbox;
 	private CustomerRef customer;
@@ -30,23 +30,23 @@ public class Conversation {
 	private UserRef closedBy;
 
 	private PersonRef createdBy;
-	
+
 	private List<String> ccList;
 	private List<String> bccList;
 	private List<String> tags;
-	
+
 	private List<LineItem> threads;
-	
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public int getFolderId() {
+	public Long getFolderId() {
 		return folderId;
 	}
 	public boolean isDraft() {
 		return isDraft;
 	}
-	public int getNumber() {
+	public Long getNumber() {
 		return number;
 	}
 	public UserRef getOwner() {
@@ -73,15 +73,15 @@ public class Conversation {
 	public Source getSource() {
 		return source;
 	}
-	
+
 	public PersonRef getCreatedBy() {
 		return createdBy;
 	}
-	
+
 	public boolean isCreatedByCustomer() {
 		return createdBy != null && createdBy instanceof CustomerRef;
 	}
-	
+
 	public Calendar getCreatedAt() {
 		return createdAt;
 	}
@@ -91,11 +91,11 @@ public class Conversation {
 	public String getClosedAt() {
 		return closedAt;
 	}
-	
+
 	public UserRef getClosedBy() {
 		return closedBy;
 	}
-	
+
 	public boolean hasCcList() {
 		return ccList != null && ccList.size() > 0;
 	}
@@ -117,7 +117,7 @@ public class Conversation {
 	public boolean hasThreads() {
 		return threads != null && threads.size() > 0;
 	}
-	
+
 	public List<LineItem> getThreads() {
 		return threads;
 	}
