+++ src/main/java/net/helpscout/api/model/ref/AbstractRef.java
@@ -1,12 +1,12 @@
 package net.helpscout.api.model.ref;
 
 public abstract class AbstractRef implements PersonRef {
-	private int id;
+	private Long id;
 	private String firstName;
 	private String lastName;
 	private String email;
-	
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
 	public String getFirstName() {
