+++ src/main/java/net/helpscout/api/model/ref/MailboxRef.java
@@ -1,13 +1,13 @@
 package net.helpscout.api.model.ref;
 
 public class MailboxRef {
-	private int id;
+	private Long id;
 	private String name;
-	
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
 	public String getName() {
 		return name;
-	}	
+	}
 }
